src/backend/requirements.txt
awsiotsdk==1.7.1
flask-login==0.5.0
flask-socketio==5.1.1
